Ensure we use backoff when request fails
Currently, the exponential backoff feature of the `Poll` class is not used since we never throw an error for a bad network request (e.g. a 404).

This is demonstrated by the fact that the frontend extension will keep hitting a 404 every 5 seconds if you disable the server extension.

As can be seen from the lumino code, we need to raise an exception to trigger the backoff code. Simply returning null doesn't do anything:
https://github.com/jupyterlab/lumino/blob/8567deefe4dd2c84b3bd22da5c8151941cd3b2db/packages/polling/src/poll.ts#L319-L329